backend/requirements.txt
# MindTube Backend Dependencies

# Core Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0

# Async & Background Tasks
asyncio-mqtt==0.16.1
arq==0.25.0
redis==5.0.1

# Database
sqlalchemy==2.0.23
alembic==1.13.1
sqlite3  # Built-in with Python

# HTTP Client & API Integration
httpx==0.25.2
aiohttp==3.9.1
requests==2.31.0

# YouTube & Media Processing
google-api-python-client==2.108.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.1.0
yt-dlp==2023.11.16

# AI/LLM Integration
openai==1.3.7
anthropic==0.7.7
tiktoken==0.5.2

# Audio Processing (for ASR)
openai-whisper==20231117
ffmpeg-python==0.2.0
pydub==0.25.1

# Text Processing
nltk==3.8.1
spacy==3.7.2
langdetect==1.0.9

# Utilities
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
tenacity==8.2.3

# Monitoring & Logging
structlog==23.2.0
sentry-sdk[fastapi]==1.38.0
prometheus-client==0.19.0

# Development & Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
httpx==0.25.2  # For testing
black==23.11.0
flake8==6.1.0
isort==5.12.0
mypy==1.7.1

# Security
cryptography==41.0.8
bcrypt==4.1.2